refactor: move MenuItem to new folder
